split sql_process_store into multiple files
